scripts/run_autonomous.py
#!/usr/bin/env python3
"""
Main script to run autonomous lunar navigation.

Usage:
    python run_autonomous.py --config config/params.yaml
"""

import argparse
import yaml
import time

# TODO: Import modules
# from sensors import StereoCamera, IMU
# from depth import StereoMatcher, DepthUncertainty
# from mapping import VoxelGrid, HeightMap
# from costmap import PerceptionAwareCostmap
# from planning import AStarPlanner, CoveragePlanner, TrajectoryGenerator
# from control import Controller
# from visualization import Visualizer


def load_config(config_path):
    """Load configuration from YAML file"""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def main(args):
    # Load configuration
    config = load_config(args.config)
    
    print("=" * 50)
    print("LUNAR NAVIGATION SYSTEM")
    print("=" * 50)
    
    # TODO: Initialize CARLA client
    # client = carla.Client('localhost', 2000)
    # world = client.get_world()
    
    # TODO: Initialize sensors
    # stereo_camera = StereoCamera(world, vehicle, config['camera'])
    # imu = IMU(world, vehicle, config['imu'])
    
    # TODO: Initialize depth
    # stereo_matcher = StereoMatcher(config['stereo'])
    # depth_uncertainty = DepthUncertainty(config['uncertainty'])
    
    # TODO: Initialize mapping
    # voxel_grid = VoxelGrid(config['mapping'])
    # height_map = HeightMap(config['mapping'])
    
    # TODO: Initialize costmap
    # costmap = PerceptionAwareCostmap(config['costmap'])
    
    # TODO: Initialize planning
    # local_planner = AStarPlanner(costmap, config['planning'])
    # coverage_planner = CoveragePlanner(config['planning'])
    # trajectory_gen = TrajectoryGenerator(config['trajectory'])
    
    # TODO: Initialize control
    # controller = Controller(config['control'])
    
    # TODO: Initialize visualization
    # visualizer = Visualizer(config['visualization'])
    
    print("All modules initialized. Starting autonomous navigation...")
    
    # Main loop
    try:
        while True:
            # TODO: Implement main loop
            # 1. Get sensor data
            # 2. Run SLAM (get pose from ORB-SLAM3)
            # 3. Compute stereo depth
            # 4. Update voxel grid
            # 5. Update costmap
            # 6. Plan path
            # 7. Generate trajectory
            # 8. Compute control
            # 9. Send to vehicle
            # 10. Update visualization
            
            time.sleep(0.1)  # 10 Hz loop
            
    except KeyboardInterrupt:
        print("\nShutting down...")
    
    # Cleanup
    # stereo_camera.destroy()
    # imu.destroy()
    # visualizer.close()


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description='Lunar Navigation System')
    parser.add_argument('--config', type=str, default='config/params.yaml',
                        help='Path to configuration file')
    args = parser.parse_args()
    main(args)
