requirements.txt
# Core
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0

# Computer Vision
opencv-python>=4.5.0

# Visualization
matplotlib>=3.4.0
open3d>=0.15.0

# CARLA (install separately)
# carla==0.9.15

# ORB-SLAM3 (build from source)
# See: https://github.com/UZ-SLAMLab/ORB_SLAM3
